+++ weblate/accounts/views.py
@@ -48,7 +48,6 @@ from django.template.loader import render_to_string
 from django.urls import reverse
 from django.utils import timezone
 from django.utils.cache import patch_response_headers
-from django.utils.crypto import get_random_string
 from django.utils.decorators import method_decorator
 from django.utils.http import urlencode
 from django.utils.translation import get_language
@@ -125,6 +124,7 @@ from weblate.utils.ratelimit import (
 )
 from weblate.utils.request import get_ip_address, get_user_agent
 from weblate.utils.stats import prefetch_stats
+from weblate.utils.token import get_token
 from weblate.utils.views import get_component, get_project
 
 CONTACT_TEMPLATE = """
@@ -666,13 +666,13 @@ class UserPage(UpdateView):
         context["page_profile"] = user.profile
         context["last_changes"] = last_changes.preload()
         context["last_changes_url"] = urlencode({"user": user.username})
-        context["user_translations"] = prefetch_stats(user_translations)
-        context["owned_projects"] = prefetch_project_flags(
+        context["page_user_translations"] = prefetch_stats(user_translations)
+        context["page_owned_projects"] = prefetch_project_flags(
             prefetch_stats(
                 user.owned_projects.filter(id__in=allowed_project_ids).order()
             )
         )
-        context["watched_projects"] = prefetch_project_flags(
+        context["page_watched_projects"] = prefetch_project_flags(
             prefetch_stats(
                 user.watched_projects.filter(id__in=allowed_project_ids).order()
             )
@@ -988,7 +988,7 @@ def reset_api_key(request):
     # Need to delete old token as key is primary key
     with transaction.atomic():
         Token.objects.filter(user=request.user).delete()
-        Token.objects.create(user=request.user, key=get_random_string(40))
+        Token.objects.create(user=request.user, key=get_token("wlu"))
 
     return redirect_profile("#api")