+++ weblate/api/tests.py
@@ -736,7 +736,7 @@ class RoleAPITest(APIBaseTest):
         self.assertEqual(response.data["count"], 2)
         self.authenticate(True)
         response = self.client.get(reverse("api:role-list"))
-        self.assertEqual(response.data["count"], 13)
+        self.assertEqual(response.data["count"], 14)
 
     def test_get_role(self):
         role = Role.objects.get(name="Access repository")
@@ -761,7 +761,7 @@ class RoleAPITest(APIBaseTest):
             format="json",
             request={"name": "Role", "permissions": ["suggestion.add", "comment.add"]},
         )
-        self.assertEqual(Role.objects.count(), 14)
+        self.assertEqual(Role.objects.count(), 15)
         self.assertEqual(Role.objects.get(name="Role").permissions.count(), 2)
 
     def test_delete(self):
@@ -772,7 +772,7 @@ class RoleAPITest(APIBaseTest):
             superuser=True,
             code=204,
         )
-        self.assertEqual(Role.objects.count(), 12)
+        self.assertEqual(Role.objects.count(), 13)
 
     def test_put(self):
         self.do_request(
@@ -1162,6 +1162,7 @@ class ProjectAPITest(APIBaseTest):
             },
         )
         self.assertEqual(response.data["repo"], repo)
+        self.assertEqual(response.data["branch"], branch)
         component = Component.objects.get(slug="c-1")
         self.assertEqual(component.repo, link_repo)
         response = self.do_request(
@@ -1181,6 +1182,7 @@ class ProjectAPITest(APIBaseTest):
             },
         )
         self.assertEqual(response.data["repo"], repo)
+        self.assertEqual(response.data["branch"], "translations")
         component = Component.objects.get(slug="c-2")
         self.assertEqual(component.repo, repo)
         response = self.do_request(
@@ -1201,6 +1203,7 @@ class ProjectAPITest(APIBaseTest):
             },
         )
         self.assertEqual(response.data["repo"], repo)
+        self.assertEqual(response.data["branch"], branch)
         component = Component.objects.get(slug="c-3")
         self.assertEqual(component.repo, repo)
 
@@ -2279,7 +2282,14 @@ class TranslationAPITest(APIBaseTest):
                 {"file": handle},
             )
         self.assertEqual(response.status_code, 400)
-        self.assertIn("detail", response.data)
+        self.assertEqual(
+            response.data,
+            {
+                "file": ErrorDetail(
+                    string="Plural forms do not match the language.", code="invalid"
+                )
+            },
+        )
 
     def test_repo_status_denied(self):
         self.do_request("api:translation-repository", self.translation_kwargs, code=403)
@@ -2853,8 +2863,8 @@ class ScreenshotAPITest(APIBaseTest):
                 code=400,
                 superuser=True,
                 data={
-                    "detail": ErrorDetail(
-                        string="Missing language_code parameter", code="parse_error"
+                    "language_code": ErrorDetail(
+                        string="This field is required.", code="invalid"
                     )
                 },
                 request={
@@ -2870,10 +2880,18 @@ class ScreenshotAPITest(APIBaseTest):
                 code=400,
                 superuser=True,
                 data={
-                    "detail": ErrorDetail(
+                    "project_slug": ErrorDetail(
                         string="Translation matching query does not exist.",
                         code="invalid",
-                    )
+                    ),
+                    "component_slug": ErrorDetail(
+                        string="Translation matching query does not exist.",
+                        code="invalid",
+                    ),
+                    "language_code": ErrorDetail(
+                        string="Translation matching query does not exist.",
+                        code="invalid",
+                    ),
                 },
                 request={
                     "name": "Test create screenshot",