+++ weblate/trans/exceptions.py
@@ -18,9 +18,16 @@
 #
 
 
-class FileParseError(Exception):
-    """Generic error for parsing."""
+class WeblateError(Exception):
+    """Base class for Weblate errors."""
+
+    def __init__(self, message=None):
+        super().__init__(message or self.__doc__)
 
 
-class PluralFormsMismatch(Exception):
+class FileParseError(WeblateError):
+    """File parse error."""
+
+
+class PluralFormsMismatch(WeblateError):
     """Plural forms do not match the language."""