+++ weblate/trans/models/change.py
@@ -17,6 +17,8 @@
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 #
 
+from datetime import datetime
+
 from django.conf import settings
 from django.db import models, transaction
 from django.db.models import Count, Q
@@ -49,9 +51,8 @@ class ChangeQuerySet(models.QuerySet):
             base = base.prefetch()
         return base.filter(action__in=Change.ACTIONS_CONTENT)
 
-    @staticmethod
-    def count_stats(days, step, dtstart, base):
-        """Count number of changes in given dataset and period grouped by step days."""
+    def count_stats(self, days: int, step: int, dtstart: datetime):
+        """Count the number of changes in a given period grouped by step days."""
         # Count number of changes
         result = []
         for _unused in range(0, days, step):
@@ -60,7 +61,7 @@ class ChangeQuerySet(models.QuerySet):
             int_end = int_start + timezone.timedelta(days=step)
 
             # Count changes
-            int_base = base.filter(timestamp__range=(int_start, int_end))
+            int_base = self.filter(timestamp__range=(int_start, int_end))
             count = int_base.aggregate(Count("id"))
 
             # Append to result
@@ -73,8 +74,8 @@ class ChangeQuerySet(models.QuerySet):
 
     def base_stats(
         self,
-        days,
-        step,
+        days: int,
+        step: int,
         project=None,
         component=None,
         translation=None,
@@ -104,11 +105,11 @@ class ChangeQuerySet(models.QuerySet):
         if user is not None:
             base = base.filter(user=user)
 
-        return self.count_stats(days, step, dtstart, base)
+        return base.count_stats(days, step, dtstart)
 
     def prefetch(self):
         """
-        Fetch related fields in a big chungs to avoid loading them individually.
+        Fetch related fields at once to avoid loading them individually.
 
         Call prefetch or prefetch_list later on paginated results to complete.
         """
@@ -139,10 +140,10 @@ class ChangeQuerySet(models.QuerySet):
         return self.preload_list(self, *args)
 
     def last_changes(self, user):
-        """Return last changes for an user.
+        """Return the most recent changes for an user.
 
-        Prefilter Changes by ACL for users and fetches related fields for last changes
-        display.
+        Filters Change objects by user permissions and fetches related fields for
+        last changes display.
         """
         if user.is_superuser:
             return self.prefetch().order()
@@ -435,7 +436,7 @@ class Change(models.Model, UserDisplayMixin):
         ),
         # Translators: Name of event in the history
         ACTION_UNLOCK: gettext_lazy(
-            "Component was automatically unlocked as alert was fixed."
+            "Fixing an alert automatically unlocked the component."
         ),
     }
 
@@ -589,9 +590,9 @@ class Change(models.Model, UserDisplayMixin):
                 )
             )
             if reason == "content changed":
-                return mark_safe(_("File %s was changed.") % filename)
+                return mark_safe(_('The "%s" file was changed.') % filename)
             if reason == "check forced":
-                return mark_safe(_("Parsing of file %s was enforced.") % filename)
+                return mark_safe(_('Parsing of the "%s" file was enforced.') % filename)
             if reason == "new file":
                 return mark_safe(_("File %s was added.") % filename)
             raise ValueError(f"Unknown reason: {reason}")
@@ -599,7 +600,7 @@ class Change(models.Model, UserDisplayMixin):
         if self.action == self.ACTION_LICENSE_CHANGE:
             not_available = pgettext("License information not available", "N/A")
             return _(
-                "License for component %(component)s was changed "
+                'The license of the "%(component)s" component was changed '
                 "from %(old)s to %(target)s."
             ) % {
                 "component": self.component,