+++ weblate/trans/views/acl.py
@@ -23,6 +23,7 @@ from django.contrib.auth.decorators import login_required
 from django.core.exceptions import PermissionDenied
 from django.http import JsonResponse
 from django.shortcuts import redirect
+from django.template.loader import render_to_string
 from django.utils import timezone
 from django.utils.translation import gettext as _
 from django.views.decorators.http import require_POST
@@ -30,9 +31,14 @@ from django.views.decorators.http import require_POST
 from weblate.accounts.models import AuditLog
 from weblate.auth.forms import InviteUserForm, send_invitation
 from weblate.auth.models import Group, User
-from weblate.trans.forms import UserBlockForm, UserManageForm
+from weblate.trans.forms import (
+    ProjectTokenCreateForm,
+    ProjectTokenDeleteForm,
+    UserBlockForm,
+    UserManageForm,
+)
 from weblate.trans.models import Change
-from weblate.trans.util import render
+from weblate.trans.util import redirect_param, render
 from weblate.utils import messages
 from weblate.utils.views import get_project, show_form_errors
 from weblate.vcs.ssh import get_key_data
@@ -252,10 +258,12 @@ def manage_access(request, project):
         {
             "object": obj,
             "project": obj,
+            "project_tokens": obj.projecttoken_set.all(),
             "groups": Group.objects.for_project(obj),
             "all_users": User.objects.for_project(obj),
             "blocked_users": obj.userblock_set.select_related("user"),
             "add_user_form": UserManageForm(),
+            "create_project_token_form": ProjectTokenCreateForm(obj),
             "block_user_form": UserBlockForm(
                 initial={"user": request.GET.get("block_user")}
             ),
@@ -263,3 +271,45 @@ def manage_access(request, project):
             "ssh_key": get_key_data(),
         },
     )
+
+
+@require_POST
+@login_required
+def delete_token(request, project):
+    """Delete project token."""
+    obj = get_project(request, project)
+
+    if not request.user.has_perm("project.edit", obj):
+        raise PermissionDenied()
+
+    form = ProjectTokenDeleteForm(obj, request.POST)
+
+    if form.is_valid():
+        form.cleaned_data["token"].delete()
+    else:
+        show_form_errors(request, form)
+
+    return redirect_param("manage-access", "#api", project=obj.slug)
+
+
+@require_POST
+@login_required
+def create_token(request, project):
+    """Create project token."""
+    obj = get_project(request, project)
+
+    if not request.user.has_perm("project.edit", obj):
+        raise PermissionDenied()
+
+    form = ProjectTokenCreateForm(obj, request.POST)
+
+    if form.is_valid():
+        token = form.save()
+        messages.info(
+            request,
+            render_to_string("trans/projecttoken-created.html", {"token": token}),
+        )
+    else:
+        show_form_errors(request, form)
+
+    return redirect_param("manage-access", "#api", project=obj.slug)